Configured About page to work with CMS
