+++ .env
@@ -2,3 +2,5 @@ PAYLOAD_SECRET=b7d8587cd3ab228d840588dcd5c43c38a71213398c645022bf03cf09f2ed16b5
 DATABASE_URI=file:./data.db
 NEXT_PUBLIC_SERVER_URL=http://localhost:3000
 NEXT_PUBLIC_GH_PAGES=false  # Toggle this to `true` ONLY during GH Pages deployment
+
+NEXT_PUBLIC_BASE_PATH=/peritext-website
